requirements.txt
PyYAML==4.2b1
boto3==1.9.88